Fix callback example

+++ examples/Callbacks/Callbacks.ino
@@ -44,6 +44,6 @@ void loop() {
 /**
  * Define callback
  */
-void toggleBacklight(uint16_t isOn) {
+void toggleBacklight(uint8_t isOn) {
     lcdAdapter.setBacklight(isOn);
 }
